requirements.txt
pytorch-ignite==0.2.1
transformers==2.1.1
tensorboardX==1.8
tensorflow==1.14.0
fastapi
scikit-learn
uvicorn
